Added my own testing tool. Update some tests and code.

+++ Questoes/Q19/solution.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-float ** create_matrix(int rows, int columns);
-void populate_matrix(float **mat, int rows, int columns);
-void free_matrix(float **mat, int rows);
-void print_matrix(float **mat, int rows, int columns);
+double ** create_matrix(int rows, int columns);
+void populate_matrix(double **mat, int rows, int columns);
+void free_matrix(double **mat, int rows);
+void print_matrix(double **mat, int rows, int columns);
 
-float ** prodesc(float **mat1, int rows1, int columns1, float **mat2, int rows2, int columns2, int *rows_output, int *columns_output);
+double ** prodesc(double **mat1, int rows1, int columns1, double **mat2, int rows2, int columns2, int *rows_output, int *columns_output);
 
 int main(void)
 {
@@ -15,19 +15,19 @@ int main(void)
 
 	int mat_rows, mat_columns;	
 	scanf("%d %d", &mat_rows, &mat_columns);
-	float **mat = create_matrix(mat_rows, mat_columns);
+	double **mat = create_matrix(mat_rows, mat_columns);
 	populate_matrix(mat, mat_rows, mat_columns);
 
 	for (int i = 1; i < n_mats; ++i)
 	{
 		int mat_mult_rows, mat_mult_columns;
 		scanf("%d %d", &mat_mult_rows, &mat_mult_columns);
-		float **mat_mult = create_matrix(mat_mult_rows, mat_mult_columns);
+		double **mat_mult = create_matrix(mat_mult_rows, mat_mult_columns);
 		if (mat == NULL || mat_mult == NULL) return 1;
 
 		populate_matrix(mat_mult, mat_mult_rows, mat_mult_columns);
 
-		float **mat_aux = prodesc(mat, mat_rows, mat_columns, mat_mult, mat_mult_rows, mat_mult_columns, &mat_rows, &mat_columns);
+		double **mat_aux = prodesc(mat, mat_rows, mat_columns, mat_mult, mat_mult_rows, mat_mult_columns, &mat_rows, &mat_columns);
 		
 		free_matrix(mat, mat_rows);
 		free_matrix(mat_mult, mat_mult_rows);
@@ -39,7 +39,7 @@ int main(void)
 }
 
 
-float ** prodesc(float **mat1, int rows1, int columns1, float **mat2, int rows2, int columns2, int *rows_output, int *columns_output)
+double ** prodesc(double **mat1, int rows1, int columns1, double **mat2, int rows2, int columns2, int *rows_output, int *columns_output)
 {
 	if (columns1 != rows2)
 	{
@@ -49,7 +49,7 @@ float ** prodesc(float **mat1, int rows1, int columns1, float **mat2, int rows2,
 	*rows_output = rows1;
 	*columns_output = columns2;
 
-	float **resulting_mat = create_matrix(*rows_output, *columns_output);
+	double **resulting_mat = create_matrix(*rows_output, *columns_output);
 
 	for (int row = 0; row < *rows_output; row++)
 	{
@@ -62,28 +62,28 @@ float ** prodesc(float **mat1, int rows1, int columns1, float **mat2, int rows2,
 	return resulting_mat;
 }
 
-float ** create_matrix(int rows, int columns) {
+double ** create_matrix(int rows, int columns) {
 	if (rows <= 0 || columns <= 0) return NULL;
 
-	float **mat = (float **) calloc(rows, sizeof(float *));
+	double **mat = (double **) calloc(rows, sizeof(double *));
 	if (mat == NULL) return NULL;
 
 	for (int i = 0; i < rows; i++) {
-		mat[i] = (float *) calloc(columns, sizeof(float));
+		mat[i] = (double *) calloc(columns, sizeof(double));
 		if (mat[i] == NULL) return NULL;
 	}
 
 	return mat;
 }
 
-void populate_matrix(float **mat, int rows, int columns)
+void populate_matrix(double **mat, int rows, int columns)
 {	
 	for (int row = 0; row < rows; row++)
 		for (int column = 0; column < columns; column++)
-			scanf("%f", &mat[row][column]);
+			scanf("%lf", &mat[row][column]);
 }
 
-void free_matrix(float **mat, int rows)
+void free_matrix(double **mat, int rows)
 {
 	for (int i = 0; i < rows; i++)
 		free(mat[i]);
@@ -92,13 +92,13 @@ void free_matrix(float **mat, int rows)
 	return;
 }
 
-void print_matrix(float **mat, int rows, int columns)
+void print_matrix(double **mat, int rows, int columns)
 {
 	for (int row = 0; row < rows; row++)
 	{
 		for (int column = 0; column < columns; column++)
 		{
-			printf("%.1f", mat[row][column]);
+			printf("%.1lf", mat[row][column]);
 			if (column < columns - 1)
 				printf(" ");
 		}
